Add docs (#19)
I added more information to the README like how to install and update
the userscripts. Furthermore, I added documentation for userscripts.

The README and the documentation for userscripts are very similar to
Mottie/GitHub-userscripts.

Markdownlint was configured such that it ignores the line length of
tables and allows the br tag. The reason is that the table in README.md
is more readable in this way, since its lines are not very long. The br
tag allows me to adjust the position of the userscript table and make
the README more visually pleasing.

See also: https://github.com/Mottie/GitHub-userscripts?rgh-link-date=2025-04-26T19%3A02%3A49Z
Closes: https://github.com/ikelax/userscripts/issues/5