Convert German date strings into English (#37)
The issue was that German date strings like "Mittwoch, 21. Mai 2025" are
not accepted by new Date() because the month Mai is written May in
English.

The function convertDate first strips of the weekday and then translates
the months in German into English. This fixes the bug.

The typos words were extended because it detected months in German as
typos. However, they are intentionally written in German.

Closes https://github.com/ikelax/userscripts/issues/36

See pull request: https://github.com/ikelax/userscripts/pull/37