Add navbar entry for HfM
The navigation bar entry is for Cafeteria Hochschule für Musik Saar
which is usually abbreviated with HfM.

I renamed the documentation file and the userscript because I added a
new canteen. Because I do not want to have to rename it everytime I add
a new canteen, I chose canteens.

Closes: https://github.com/ikelax/userscripts/issues/35