Add tests for Mensaar Show Next Day (#52)
I extracted the code for computing the closing time to the function
getClosingTime in utils/getClosingTime.js and wrote tests for it.

I added "type": "module" to package.json since there were warning when
running npm test because utils/getClosingTime.js is an ES module.

eslint-plugin-userscripts needed to be restricted to userscripts/. It
was throwing false lint errors in tests/ and utils/ because the js files
did not have userscript metadata.

Closes: https://github.com/ikelax/userscripts/issues/40
See pull request: https://github.com/ikelax/userscripts/pull/52