Refactor Mensaar Show Next Day
I refactored switchToNextDay with utils/getClosingTime.js.
getClosingTime was implemented in #52 and is basically just the code
from switchToNextDay extracted into a function.

waitForKeyElements was removed because it is also possible to just load
the code with the @require keyword. This URL is from its README. The
advantage of this approach is that it is not necessary any more to
manually update the code.

Also, I bumped the minor version since we forgot this in #47.

See also: https://github.com/ikelax/userscripts/pull/47
See also: https://github.com/ikelax/userscripts/pull/52
See also: https://github.com/CoeJoder/waitForKeyElements.js